requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.23.0
opencv-python>=4.8.0
numpy>=1.24.0
werkzeug>=3.0.0
